Add Ndef write/format functionality for Mifare 1k/4K
Test: Format and write ndef data with Mifare-1K/4K tag

+++ src/nfc/include/rw_int.h
@@ -608,7 +608,7 @@ typedef struct {
 /* Mifare Classic RW Control Block */
 
 typedef struct {
-  uint8_t block;
+  uint16_t block;
   bool auth;
 } tRW_MFC_BLOCK;
